+++ CHANGELOG.md
@@ -1,14 +1,43 @@
 # Changelog
 
-All notable changes to Cosmos Cove Control Panel will be documented below.
+
+## [0.3.6] - 2025-05-10
+
+### Added
+- Custom and Maincloud server mode can now be used.
+
+### Fixed & Changed
+- Database Window now lets you DEL and CLR in all types of tables.
+- Reducer Window now properly let's you call reducers with params.
+
+
+
+## [0.3.5] - 2025-05-10
+
+### New Features
+- Initial support for Custom and Maincloud server.
+
+### Fixed & Changed
+- UI for view logs, browse database and run reducer is now clearer and with icons.
+- Refactored scripts to be more focused and less all in one.
+
+
+
+## [0.3.4] - 2025-05-12
+
+### New Features
+- Prepared for Custom and Maincloud server mode.
+
+### Fixed & Changed
+- UI fixes.
+
+
 
 ## [0.3.3] - 2025-05-10
 
-### Added
+### New Features
 - Initial release of the alpha version.
 
-### Fixed
+### Fixed & Changed
 - WSL1 will now install correctly.
-
-### Changed
 - Everything.